+++ src/AlgorithmInterface.php
@@ -1,7 +1,0 @@
-<?php
-namespace Michaels\Midas;
-
-interface AlgorithmInterface
-{
-    public function run($data, array $params = null);
-}