+++ src/Commands/CommandNotFoundException.php
@@ -1,0 +1,7 @@
+<?php
+namespace Michaels\Midas\Commands;
+
+use Michaels\Midas\Algorithms\AlgorithmNotFoundException;
+
+class CommandNotFoundException extends AlgorithmNotFoundException {}
+