+++ src/Data/RefinedData.php
@@ -1,0 +1,13 @@
+<?php
+namespace Michaels\Midas\Data;
+
+use Illuminate\Support\Collection;
+
+/**
+ * Class RefinedData
+ * @package Michaels\Midas\Data
+ */
+class RefinedData extends Collection
+{
+    // To be extended
+}