+++ src/Exceptions/CommandNotFoundException.php
@@ -1,7 +1,0 @@
-<?php
-namespace Michaels\Midas\Exceptions;
-
-use Exception;
-
-class CommandNotFoundException extends AlgorithmNotFoundException {}
-